delibot/requirements.txt
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
packaging==21.3
Pillow==9.2.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytesseract==0.3.10
yarl==1.8.1
